requirements.txt
Django==2.2.5
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-configurations==2.1
Pillow==6.2.1

